# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mi Tienda Online API",
        "description": "API REST para gestión de tienda Vue + Laravel",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/productos": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Obtener el catálogo completo de productos",
                "operationId": "2af0fdbcd7c972b8f9bd45123cd57cfb",
                "responses": {
                    "200": {
                        "description": "Catálogo recuperado con éxito"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Crear un nuevo producto",
                "operationId": "434c8599d4fe9aaa5aee59a5be0cf3fd",
                "responses": {
                    "201": {
                        "description": "Producto creado con éxito"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No tienes los permisos requeridos"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/productos/{id}": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Obtener el detalle de un producto",
                "operationId": "7f8c025cedd4418ed0240a21f7f9402a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto encontrado con éxito"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Productos"
                ],
                "summary": "Actualizar un producto existente",
                "operationId": "82ed71c67980b16764a81d295405c031",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto actualizado con éxito"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Productos"
                ],
                "summary": "Eliminar un producto",
                "operationId": "075714328668dc259b9f6efb188ab787",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto eliminado con éxito"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "sanctum-token",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Productos",
            "description": "Productos"
        }
    ]
}